+++ CHANGELOG.md
@@ -13,6 +13,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make data tab, credential details, activity and add-data screens schema-independent
 - Add certificate management in developer mode
 
+## [7.13.5] - 2026-04-22
+### Changed
+- Remove unsupported countries from sms issuance phone entry screen
+- Add some missing country codes in sms issuance phone enter screen
+- Automatically convert email input to lowercase in email issuance screen
+- Update to vcmrtd v3.8.0 to fix German passport reading issues
+
 ## [7.13.4] - 2026-04-01
 ### Changed
 - Add Tesseract4Android for F-Droid OCR 
@@ -602,6 +609,7 @@ This release only includes iOS changes.
 - Log screen now shows all log items
 - Various bug fixes
 
+[7.13.5]: https://github.com/privacybydesign/irmamobile/compare/v7.13.4...v7.13.5
 [7.13.4]: https://github.com/privacybydesign/irmamobile/compare/v7.13.3...v7.13.4
 [7.13.3]: https://github.com/privacybydesign/irmamobile/compare/v7.13.2...v7.13.3
 [7.13.2]: https://github.com/privacybydesign/irmamobile/compare/v7.13.1...v7.13.2